Add files via upload
The context
A company that is active in Big Data and Data Science wants to hire data scientists among people who successfully pass some courses which conduct by the company. Many people signup for their training. The company wants to know which of these candidates are really want to work for the company after training or looking for new employment because it helps to reduce the cost and time as well as the quality of training or planning the courses and categorization of candidates. Information related to demographics, education, experience. 
This dataset designed to understand the factors that lead a person to leave their current job for HR researches too.
Features
•	enrollee_id : Unique ID for candidate
•	city: City code
•	city_ development _index : Developement index of the city (scaled)
•	gender: Gender of candidate
•	relevent_experience: Relevant experience of candidate
•	enrolled_university: Type of University course enrolled if any
•	education_level: Education level of candidate
•	major_discipline :Education major discipline of candidate
•	experience: Candidate total experience in years
•	company_size: No of employees in current employer's company
•	company_type : Type of current employer
•	lastnewjob: Difference in years between previous job and current job
•	training_hours: training hours completed
•	target: 0 – Not looking for job change, 1 – Looking for a job change
Inspiration
Learn why candidates quit the training program. Are the reasons come from company treatment or from other reasons. 
The result
A large of candidates who decide to look for a new job are just graduated, no experience and are male.